feat: implement VerifyVCHashOnChain method and update README; remove redundant verification strategies
