+++ credential/vccontract/README.md
@@ -2,17 +2,17 @@
 
 `vccontract` is a lightweight, **read-only** client for the Credential Registry
 smart contract. It lets a holder or any third party verify that a Verifiable
-Credential (VC) hash is anchored on-chain, without a private key and without
+Credential (VC) hash is anchored on-chain — without a private key and without
 spending gas (every call is an `eth_call`).
 
 ## How anchoring works
 
 The issuer service groups VC hashes per issuer into Merkle trees and publishes
 each tree's **root** on-chain (one root per `issuer` + `treeIndex`). A VC hash is
-a **leaf**. Membership is proven with a Merkle proof — the ordered list of
-sibling hashes that fold up to the root. Nodes are combined with
-`keccak256(sort(a, b))` (sorted-pair hashing), matching the tree construction in
-the issuer service.
+a **leaf**. Membership is proven with a Merkle **proof** — the ordered list of
+sibling hashes that fold up to the root. Verification (folding the proof and
+comparing to the stored root) is done by the contract itself, so this package
+only has to pass the proof through.
 
 ## Inputs
 
@@ -44,21 +44,34 @@ req := &vccontract.VerifyRequest{
     Proof:         []string{"0x...", "0x..."},
 }
 
-ok, err := registry.VerifyOnChain(context.Background(), req)
+ok, err := registry.VerifyVCHashOnChain(context.Background(), req)
 ```
 
-## Two verification strategies
+`VerifyVCHashOnChain` calls the contract's `verifyVC(issuer, treeIndex, leaf, proof)`
+view function. A `true` result means the VC hash is anchored in the issuer's tree;
+a `false` result (with a `nil` error) means the proof does not validate. A non-nil
+error means the call itself failed — bad input, RPC error, or the tree does not
+exist.
 
-- **`VerifyOnChain`** — calls the contract's `verifyVC(issuer, treeIndex, leaf, proof)`
-  view function. The contract does the folding, so the result is exactly what the
-  chain accepts.
-- **`VerifyByCompareRoot`** — fetches the root via `getTreeRoot(issuer, treeIndex)` and
-  re-computes it locally from `leaf` + `proof`, then compares. This is an
-  independent check that does not rely on the contract's verify logic — useful for
-  cross-validating `VerifyOnChain`.
+## API
 
-For a well-formed proof both return the same result. A `false` result (with a
-`nil` error) means the proof does not validate; a non-nil error means the call
-itself failed (bad input, RPC error, or the tree does not exist).
+- `NewCredentialRegistry(rpcURL, contractAddress string) (*CredentialRegistry, error)` —
+  connect to the chain (RPC connection is required).
+- `(*CredentialRegistry) VerifyVCHashOnChain(ctx, *VerifyRequest) (bool, error)` —
+  verify a VC hash against its on-chain tree.
+- `(*CredentialRegistry) GetTreeRoot(ctx, issuer, treeIndex) ([32]byte, error)` —
+  read the anchored Merkle root (zero value = no such tree).
+- `(*CredentialRegistry) HasTree(ctx, issuer, treeIndex) (bool, error)` —
+  whether the issuer has an anchored tree at that index.
+- `(*CredentialRegistry) Close()` — release the RPC connection.
 
-Helper reads: `GetTreeRoot` and `TreeExists`.
+Reuse a single `CredentialRegistry` across calls (it holds a live, concurrency-safe
+RPC client) and `Close()` it on shutdown rather than creating one per request.
+
+## Example
+
+A runnable example against the testnet lives in `example/`:
+
+```
+go run ./credential/vccontract/example
+```
